work in progress (divide into diff files + rename vars)
