+++ dpdk/prefilter/metadata-l4-udp.c
@@ -1,0 +1,50 @@
+/* Copyright (C) 2022 Open Information Security Foundation
+*
+* You can copy, redistribute or modify this Program under the terms of
+* the GNU General Public License version 2 as published by the Free
+* Software Foundation.
+*
+* This program is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License
+* version 2 along with this program; if not, write to the Free Software
+* Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
+* 02110-1301, USA.
+ */
+
+/**
+* \file
+*
+* \author Andrei Shchapaniak <xshcha00@vutbr.cz>
+*
+ */
+
+#include "metadata-l4-udp.h"
+#include "metadata.h"
+
+int MetadataDecodePacketUDP(metadata_t *meta_data, uint16_t len) {
+    uint16_t udp_raw_len;
+
+    if (unlikely(len < UDP_HEADER_LEN)) {
+        return UDP_HLEN_TOO_SMALL;
+    }
+
+    udp_raw_len = rte_be_to_cpu_16(meta_data->udp_hdr->dgram_len);
+    if (unlikely(len < udp_raw_len)) {
+        return UDP_PKT_TOO_SMALL;
+    }
+
+    if (unlikely(len != udp_raw_len)) {
+        return UDP_HLEN_INVALID;
+    }
+
+    meta_data->src_port = rte_be_to_cpu_16(meta_data->udp_hdr->src_port);
+    meta_data->dst_port = rte_be_to_cpu_16(meta_data->udp_hdr->dst_port);
+    meta_data->payload_len = len - UDP_HEADER_LEN;
+    meta_data->l4_len = UDP_HEADER_LEN;
+
+    return 0;
+}
